requirements.txt
fastapi
uvicorn
pydantic
sqlmodel

numpy
requests
beautifulsoup4
orjson

httpx
chromadb

openai
python-dotenv